% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_edge_comm.R
\name{community_overlap}
\alias{community_overlap}
\title{Compare how much communities overlap with each other}
\usage{
community_overlap(communities)
}
\description{
Compare how much communities overlap with each other
}
