% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{dist_nested}
\alias{dist_nested}
\title{Create a distance matrix based on nestedness: distance(i, j) = 1 - nest(i, j)}
\usage{
dist_nested(g)
}
\arguments{
\item{g}{The input \code{igraph} graph object}
}
\value{
A matrix of size \code{igraph::vcount(g)} x \code{igraph::vcount(g)}, where the (i, j)-th entry contains the distance between the i-th and j-th vertices
}
\description{
Create a distance matrix based on nestedness: distance(i, j) = 1 - nest(i, j)
}
