% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_edge_comm.R
\name{edge_comm_postprocess}
\alias{edge_comm_postprocess}
\title{Perform edge community post-processing by removing communities that are
just subsets of other communities.}
\usage{
edge_comm_postprocess(edge_communities, check_only = NULL, remap = TRUE)
}
\arguments{
\item{edge_communities}{The edge communities list of the algorithm}

\item{check_only}{Optionally a vector with colours to specifically check.
If \code{NULL}, all colours will be checked, otherwise only the specified ones.}

\item{remap}{Whether to remap communities after cleaning to the range \code{1:n}}
}
\description{
Perform edge community post-processing by removing communities that are
just subsets of other communities.
}
