% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_edge_comm.R
\name{edges_per_community}
\alias{edges_per_community}
\title{Returns the list of communities with the edges or nodes they contain}
\usage{
edges_per_community(communities)
}
\description{
Returns the list of communities with the edges or nodes they contain
}
