% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_edge_comm.R
\name{find_excluded_colors}
\alias{find_excluded_colors}
\title{Find colours that cannot be assigned to both \code{node_a} and \code{node_b} because
that would create a 2K_2.}
\usage{
find_excluded_colors(
  node_a,
  node_b,
  edges_a,
  edges_b,
  node_communities,
  edge_communities,
  neighs1,
  neighs2
)
}
\value{
A \code{vector} of exluded colours
}
\description{
Find colours that cannot be assigned to both \code{node_a} and \code{node_b} because
that would create a 2K_2.
}
