% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_node_comm.R
\name{find_maximal_sets}
\alias{find_maximal_sets}
\title{Find maximal sets from the results of \code{nested_node_comm}}
\usage{
find_maximal_sets(list_of_lk)
}
\arguments{
\item{list_of_lk}{The result list of \code{nested_node_comm}}
}
\value{
A list of maximal sets
}
\description{
Find maximal sets from the results of \code{nested_node_comm}
}
