% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_edge_comm.R
\name{find_star_communities}
\alias{find_star_communities}
\title{Find communities that are stars}
\usage{
find_star_communities(g, edge_communities)
}
\value{
A binary mask of communities that are stars
}
\description{
Find communities that are stars
}
