% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_graphs.R
\name{generate_benchmark_from_graph}
\alias{generate_benchmark_from_graph}
\title{Expand a directed base graph (tree or forest) into a bipartite graph with overlapping nested structure. An \code{a -> b} directed edge means the neighborhood of \code{a} is a subset of the neighborhood of \code{b} (i.e. \code{a} will be nested with \code{b}).
It is suggested to use \code{generate_benchmark_random} instead of directly calling this function.}
\usage{
generate_benchmark_from_graph(base_g)
}
\arguments{
\item{base_g}{An \code{igraph} graph object. Should be directed and acyclic. Does not need to have only one component.}
}
\value{
A bipartite graph where the nodes of \code{base_g} are nested in the way their edges are directed. This can create overlapping nested communities.
}
\description{
Expand a directed base graph (tree or forest) into a bipartite graph with overlapping nested structure. An \code{a -> b} directed edge means the neighborhood of \code{a} is a subset of the neighborhood of \code{b} (i.e. \code{a} will be nested with \code{b}).
It is suggested to use \code{generate_benchmark_random} instead of directly calling this function.
}
