% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_graphs.R
\name{generate_benchmark_random}
\alias{generate_benchmark_random}
\title{Generates a random tree/forest (depending on the value of \code{blocks}) and calls \code{generate_benchmark_from_graph} on that.}
\usage{
generate_benchmark_random(
  n,
  seed = as.integer(runif(1, max = 10000)),
  blocks = length(n)
)
}
\arguments{
\item{n}{A vector of integer numbers containing the number of vertices each component has.}

\item{seed}{A random seed to set before generating the random base forest.}

\item{blocks}{The number of components to generate. Defaults to \code{length(n)}.}
}
\description{
Generates a random tree/forest (depending on the value of \code{blocks}) and calls \code{generate_benchmark_from_graph} on that.
}
