% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_edge_comm.R
\name{get_node_communities}
\alias{get_node_communities}
\title{turn edge-communities into node communities:
a node belongs to all communities its incident edges belong to
IDs of incident edges.}
\usage{
get_node_communities(n, edge_communities, node_incident_edges)
}
\arguments{
\item{n}{The number of vertices the graph has}

\item{edge_communities}{The edge_communities list of the algorithm}

\item{node_incident_edges}{List of which node is incident with which edges
(can be obtained using \code{igraph::incident_edges()})}
}
\description{
turn edge-communities into node communities:
a node belongs to all communities its incident edges belong to
IDs of incident edges.
}
