% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{girvan_newman_nestedness}
\alias{girvan_newman_nestedness}
\title{Top-down hierarchical nestedness clustering using the Girvan-Newman algorithm}
\usage{
girvan_newman_nestedness(nested_g, method = c("nested", "full"))
}
\arguments{
\item{nested_g}{The input nestedness \code{igraph} graph object obtained using \code{nestedness_graph()}}

\item{method}{Either \code{full} (closer to the original Girvan-Newman implementation), or \code{nested} (faster, does not calculate edge betweenness, only uses nestedness values from the edges)}
}
\value{
A list of vectors of length \code{igraph::vcount(g)}, where the i-th vector contains the membership of the i-th vertex. Each vector corresponds to a level (step) of hierarchical clustering.
}
\description{
Top-down hierarchical nestedness clustering using the Girvan-Newman algorithm
}
