% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{girvan_newman_step_full}
\alias{girvan_newman_step_full}
\title{Performs an iteration of the top-down hierarchical nested clustering algorithm using the "full" version}
\usage{
girvan_newman_step_full(g)
}
\arguments{
\item{g}{The input nestedness graph object}
}
\description{
Performs an iteration of the top-down hierarchical nested clustering algorithm using the "full" version
}
