% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{girvan_newman_step_nested}
\alias{girvan_newman_step_nested}
\title{Performs an iteration of the top-down hierarchical nested clustering algorithm}
\usage{
girvan_newman_step_nested(g)
}
\arguments{
\item{g}{The input nestedness graph object}

\item{method}{The same as \code{method} in \code{girvan_newman_nestedness}}
}
\description{
Performs an iteration of the top-down hierarchical nested clustering algorithm
}
