% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{hclust_to_membershiplist}
\alias{hclust_to_membershiplist}
\title{Converts a \code{hclust} results object into a list of cluster levels (one level per unique merge distance)}
\usage{
hclust_to_membershiplist(hclust_res)
}
\description{
Converts a \code{hclust} results object into a list of cluster levels (one level per unique merge distance)
}
