% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_node_comm.R
\name{is_in_list}
\alias{is_in_list}
\title{Shorthand function to check if a set part of any of the list of sets}
\usage{
is_in_list(set, list_of_sets)
}
\description{
Shorthand function to check if a set part of any of the list of sets
}
