% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{mean_partition_nestedness}
\alias{mean_partition_nestedness}
\title{Calculates the mean nestedness for each partition of each level in a result list}
\usage{
mean_partition_nestedness(g, results_list)
}
\arguments{
\item{g}{The original graph}

\item{results_list}{The result list obtained from \code{girvan_newman_nestedness}}
}
\value{
A list of vectors (each of length \code{k} - the number of partitions) of length \code{igraph::vcount(g)}, where the i-th vector contains the mean nestedness of the i-th level's partitions.
}
\description{
Calculates the mean nestedness for each partition of each level in a result list
}
