% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighs2.R
\name{neighborhood_2}
\alias{neighborhood_2}
\title{Return vertices that are reachable from \code{nodes} in two steps.
As opposed to \code{igraph::neighborhood(..., order = 2)} this functiondoes not return immediate neighbors if they are not reachable in two steps.}
\usage{
neighborhood_2(g, nodes = igraph::V(g))
}
\arguments{
\item{g}{An \code{igraph} graph object}

\item{nodes}{(Optional) The set of nodes for which the list of 2-neighbors will be returned.}
}
\description{
Return vertices that are reachable from \code{nodes} in two steps.
As opposed to \code{igraph::neighborhood(..., order = 2)} this functiondoes not return immediate neighbors if they are not reachable in two steps.
}
