% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nested_direction}
\alias{nested_direction}
\title{Checks if there is a $2K_2$ between two vertices in graph \code{g}
by calculating $|N(v1) cap N(v2)| / min(|N(v1)|, |N(v2)|)$.
This version uses only the neighborhoods of said vertices, optionally
excluding \code{v1} and \code{v2} from each other's neighborhoods if \code{v1} or
\code{v2} are non-negative.}
\usage{
nested_direction(n1, n2, v1, v2)
}
\arguments{
\item{n1}{The neighborhood of \code{v1}}

\item{n2}{The neighborhood of \code{v2}}

\item{v1}{(Optional) The first vertex of the pair}

\item{v2}{(Optional) The second vertex of the pair}
}
\value{
The direction of nestedness between \code{v1} and \code{v2}. Returns 1 if \code{n1} contains \code{n2}, -1 if \code{n2} contains \code{n1} and 0 if the two nodes are not nested.
}
\description{
Checks if there is a $2K_2$ between two vertices in graph \code{g}
by calculating $|N(v1) cap N(v2)| / min(|N(v1)|, |N(v2)|)$.
This version uses only the neighborhoods of said vertices, optionally
excluding \code{v1} and \code{v2} from each other's neighborhoods if \code{v1} or
\code{v2} are non-negative.
}
