% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_node_comm.R
\name{nested_node_comm}
\alias{nested_node_comm}
\title{Build nested communities in a hierarchical fashion}
\usage{
nested_node_comm(g, use.names = FALSE)
}
\arguments{
\item{g}{The input graph (an \code{igraph} graph object)}
}
\value{
A list of \code{data.table} objects where list item \code{k} corresponds to nested sets of length \code{k}
}
\description{
Build nested communities in a hierarchical fashion
}
