% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{partition_dataframe}
\alias{partition_dataframe}
\title{Compute performance properties for each clustering level in a result list}
\usage{
partition_dataframe(g, results_list)
}
\arguments{
\item{g}{The original graph}

\item{results_list}{The result list returned by \code{girvan_newman_nestedness}}
}
\value{
A data frame with properties of the clustering levels (number of clusters, mean nestedness, etc.)
}
\description{
Compute performance properties for each clustering level in a result list
}
