% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust.R
\name{stat_prefilter}
\alias{stat_prefilter}
\title{(Experimental) Perform statistical pre-filtering on a graph, removing edges that are unlikely to be nested}
\usage{
stat_prefilter(g, alpha = 0.05)
}
\arguments{
\item{g}{The input \code{igraph} graph object}
}
\description{
(Experimental) Perform statistical pre-filtering on a graph, removing edges that are unlikely to be nested
}
